backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
boto3==1.42.57
botocore==1.42.57
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.133.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
jmespath==1.1.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.22
resend==2.19.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wrapt==2.1.2
